graph: support choosing specific plugins (#1917)
This is the dual to #1858; it allows the cli/graph command to choose to
load only specific plugins.

This is helpful if, for example, you load an instance where you don't
have the credentials to re-load one of the plugins, but want to update
the other plugins. Then you can load and graph the plugins for which you
do have data, but leave the graph for the inaccessible plugin
un-touched.

Test plan: Manual testing, as with #1858 and for similar rationale.