cli/load should fail if any plugin fails to load (#1922)
Previously, if any individual plugin failed to load, the load command
would print an error, but return exit code 0.

As of this commit, the load command returns 1 if any plugin(s) fail to
load, and prints the name(s) of the offending plugins.

Test plan: CLI doesn't have automated testing, so induce failure in an
instance by e.g. setting the SOURCECRED_GITHUB_TOKEN="" in an instance
which uses the GitHub plugin.

Observe that the full error message is still printed to console, and
that the command returns exit code 1, and that if more than one plugins
fail that the message is printed nicely showing all the failed plugins.

Also observe that load's behavior on successful loads is unchanged.