add `sourcecred go` command (#1923)
The `go` command performs a full load, running the following commands in
sequence:
- load
- graph
- score

We don't include the `site` command, on the premise that the user might
have reasons to not include the site (e.g. they just want to use the
ouptut data for analysis), so we shouldn't add it for them. Also, if
they installed the site in the past, it will still be setup.

However, this does mean that downstream users who DO want a site should
be sure to run `sourcecred site` manually whenever they update
SourceCred versions.

Test plan: In a SC instance, run `sourcecred go` and observe that:
- in the success case, it runs the three commands in sequence
- in the failure case, it prints a message showing which command failed
(you can induce a failure by removing the cache, or unsetting
GitHub/Discord tokens).