Publish SourceCred with executable binary (#1920)
This updates our build so that we include a shebang at the top of
`bin/sourcecred.js`, mark it executable, and make it the binary entry
point for the package.

As a result, dependents can install the `sourcecred` command via npm and
conveniently use it. Also, local invocation no longer needs to be
preceded with `node`, i.e. `./bin/sourcecred.js ARGS` is sufficient.

Test plan:

Verify that `yarn test --full` passes; and that
`./bin/sourcecred.js --version` prints the version.

Also, see sourcecred/example-instance#1, which uses a prerelease package
published from this commit.