GitHub: Mint cred only for merged PRs
This commit modifies the GitHub createGraph method so that it now
returns a WeightedGraph, and that WeightedGraph sets weight 0 for any
pull request which hasn't been merged.

This improves Cred robustness, since it's easy to (non-maliciously)
create a bunch of unmerged PRs, but getting them merged is a signal of
quality.

Test plan: Unit tests added.