add JsonLog for serializing logs of JSON objects (#1931)
* add JsonLog for serializing logs of JSON objects

This commit adds the JsonLog module, which has a JsonLog class for
managing an append-only log consisting entirely of JSON objects.

In addition to this basic functionality, it provides a toString() method
which creates a string in which each log item is on its own line. This
is a much better format for storing an append-only log inside of a Git
respository, as we will avoid churning the whole file's diff when adding
additional records.

Acknowledgements to @wchargin for suggesting the seralization format,
and to @Beanow for writing [another version] of this module.

[another version]: https://github.com/sourcecred/sourcecred/commit/42809547e5d08777b9f1cd1286687566ab669d86

Test plan:
Unit tests added; `yarn test` passes.

* Remove support for writing log comments

This commit removes support for writing comments in the output logs.
This is at @wchargin's suggestion, since the putting comments blocks
e.g. `jq` from processing them.

Conservatively, I've left support for reading files with comments.

Test plan: Existing unit tests pass.

* Take @Beanow's suggestion to write plain JSON

This changes the JsonLog so that it writes out regular Json, execept
with a custom whitespacing format which puts every array item on its own
line. This way we get efficiently-packable git diffs, but also can just
use regular JSON parsers.

Test plan: Unit tests and docstrings updated.