Represent Grain as Strings at rest (#1938)
This commit modifies the grain module so that we store Grain values as
strings at rest (the full string representation of the BigInt). This
allows us to get rid of a number of flow hacks around Grain usage, and
allows us to include Grain directly in JSON objects without needing to
worry about changing serialization.

The downside is that we now need to use methods for all the basic
operations on Grain (e.g. Grain.add, Grain.sub), and that there's more
performance overhead when using Grain.

This is an alternative to #1936, which I discussed with @wchargin and we
decided was too hacky.

Test plan: All unit tests updated; `yarn test` passes.