ledger: add acounts and grain balances (#1946)
This commit updates the Ledger so that we now support tracking and
updating users' Grain balances, and lifetime paid amounts.

We don't yet add any events which actually provide Grain, so it's
impossible to use the public API to get any non-empty balances in the
Ledger. Instead, we just add the underlying logic for tracking Grain
amounts. Most notably, we handle the edge case where a user is
un-linking an alias, and we need to ensure that we can record
that the alias has already been paid, so that attackers
can't use un-linking to double-dip against the BALANCED payout strategy.

Test plan: Comprehensive unit tests added; `yarn test` passes; coverage
still at 100%.