ledger: support distributing Grain (#1947)
This commit adds `ledger.distributeGrain`, which uses a GrainPolicy and
a CredHistory to distribute Grain to everyone who has Cred.

It's basically a wrapper around calling `createGrainAllocation` and
canonicalizing it into the ledger's log. The resultant distribution has
a `credTimestamp` which is distinct from the event timestamp, since it
tracks the timestamp of the interval against which we distributed Cred,
rather than the timestamp when we actually created the event.

Added thorough unit testing, particularly validating the edge case where
we have a BALANCED policy and have unlinked aliases with retroactive
paid allocations.

Test plan: `yarn test`; unit tests are thorough.