ledger enforces timestamp ordering
This commit modifies the ledger so that it enforces that actions are
always in timestamp order. Though we don't yet need this to be true,
I feel like adding this invariant will make it easier to reason about
the ledger in the future... and it's just a sensible property.

Test plan: The ledger now throws if there are out-of-order events, and
unit tests verify this.