Add a parser to the address module (#1956)
Use of parsers is becoming widespread in the codebase, and it's annoying
to not have a default parser written for one of our most fundamental
data types. (In particular, sinc we regularly use addresses as
identifiers, they are ubiquitous in serialized data.)

@wchargin advised against this because it makes the address module less
lightweight from a dependency perspective, I think the convenience is
worth it.

Test plan: Unit tests added; `yarn test` passes; grepped for existing
usage to replace it. Trivial change really.