Dicsord: update snapshot fetcher (#1948)
This change updates the snapshot fetcher to fetch/save
data in the nested fashion in which we fetch it.

For example, it would be convenient to return message
reaction data for a message we have stored as a snapshot,
so that we can test fetching/saving the message and reaction data
together. Currently, we are fetching reaction data for a
message that isn't saved as a snapshot.

Test plan: No tests added. Unit tests added. This change will
be convenient for end to end testing in the Discord module.

Co-authored-by: Brian Litwin <blitwin@horizoninvestments.com>