Simplify Grain Allocation module (#1957)
The GrainAllocation module was pretty messy, in part because we weren't
sure about the scope when we were writing it. Therefore, we included
some unncessary concepts, like putting version numbers in many of the
module's data types. In reality, since the Ledger is responsible for
serialization, all concerns around versioning should live in its Action
data types; we don't need to think about versions here.

Also, while writing the Ledger I realized that we want a first class
concept of "Strategy + Budget", and need a first class concept of
"Allocations + timestamp of the Cred used to generate them". Thus, I've
re-written the allocation module to expose the Policy and Distribution
types, respectively.

The overall result is a simpler module, and pulling some logic out of
the Grain Ledger and into the Allocations. We could probably go further
(see notes in comments) but I think this is fine for now.

Along the way, I accidentally found an instance of a nasty bug that I've
suspected exists: we can break the BALANCED distribution method so that
it tries to distribute more Grain than is in its budget. I left a
skipped test case as a repro, will address this in another commit.

Also: I added a parser to the module.

Test plan: Unit tests updated, run `yarn test`