Add stress-tested algo for splitting Grain budgets
In #1957, I encountered a serious bug in our Grain distribution logic
that could result in distributing more Grain than was budgetted. This
was kind of unsurprising; doing arithmetic that mixes floats and bigints
is nontrivial.

This commit adds a stress-tested method, `splitBudget`, which splits a
fixed Grain budget based on floating-point scores. It
maintains the invariant that the amount distributed is always exactly
equal to the budget, even though this may result in slightly unfair
splits sometimes. (The unfairness comes about at scales like 10^-15
Grain, so I'm not too worried about it.)

The method is stress-tested by increasingly crazily example splits, so
we can have some confidence that it will satisfy its contract in
production. It has some routines that act to hide errors (e.g. letting
the last address absorb any remaining budget). So that we can get
confidence that we're not hiding large discrepancies, I've left some
commented-out "error sentinels" in the code, which log the size of the
discrepancies. These may be uncommented to see the size of the errors.

Along the way, I improved the accuracy of the `multiplyFloat` method so
that in the special case of multiplying by `1`, it returns the input
Grain value exactly.

Test plan: Inspect included unit tests, perhaps uncommenting some of the
"error sentinels".