Reify a PluginId type (#1961)
This commit adds a PluginId type, which is an opaque subtype of string
with an `owner/name` structure, modeled after GitHub repoIds.

This gives us type safety so that we always validate potential PluginIds
early, and error clearly if they are not valid. Yay!

I put the PluginId type in the `api` folder; I'm planning to migrate
most of the CLI logic into `src/api`. Also, I took an opportunity to
clean up the `InstanceConfig` parser, which was written before we got a
good feel for how to use the parser module.

Test plan: Unit tests added, `yarn flow` passes, and for sanity I
verified that the `load` and `graph` commands work properly when given
an explicit PluginId.