Start the DiskInstance class
This class is going to abstract over interacting with SourceCred
instances that are on disk. This commit just starts the class with
extremely simple scope: it loads the appropriate plugins for a SC
instance, based on the contents of `sourcecred.json`, and it errors
if the `sourcecred.json` file can't be found, or if specified plugins
are unavailable. Since this needs to be async, it happens in a static
method, called `instantiate`.

I took care with this method to emit very clear errors when it fails,
communicating exactly what went wrong (instance dir wasn't a directory,
or there was no sourcecred.json file, or the sourcecred.json file was
invalid, etc). This way we will give the user helpful feedback when
there's something basic broken in their instance setup.

Test plan: See included tests; `yarn test` passes.