Ensure Grain allocations match their budget
This commit updates the grainAllocation module so that distributions
always distribute their budgets precisely, or fail. It does this by
incorporating the new stress-tested `splitBudget` method introduced over
in #1959, and modifying the contract so that we fail rather than
returning an empty allocation in the case when totalCred sums to 0.

Test plan: `yarn test` passes, and we are now using a stress-tested core
algorithm for the distribution, rather than some ad-hoc but undertested
math.