Rebrand `TestTaskReporter` to `SilentTaskReporter` (#1967)
We will want a TaskReporter which doesn't have any side-effects,
so we can use it as a default option in situations (such as loading
plugin data) that expect one to be available. This default
TaskReporter should not have any side-effects. Happily, we have a
candidate available: the `TestTaskReporter` already exists.

This commit renames it, and adds an additional test showing that the
reporter does not emit anything to console.

Test plan: `yarn test` passing is sufficient.