addAlias emits explicit transfer events
This commit implements a great [review suggestion] by @topocount. We
change the `addAlias` method so that rather than implicitly transferring
an alias's Grain balance to the canonical user, it does so via an
explicit transfer action. This will make the transfer much easier to audit
and reverse in a case of fraudulent linking.

This has two other benefits:
1. It means we will now have an invariant that an accounts Grain balance
is equal to the sum of its transfers and its distributions. Very nice
for auditing.

2. While implementing it, I spotted a nasty bug where instead of
transferring an alias's balance to the user, it actually doubled the
user's balance! This bug was hidden because in the test case, I so
happened to give the user and the alias the same amount, so there wasn't
an observable difference.

Since nothing guarantees transactional semantics other than very careful
implementation of the ledger, I have been quite careful when
implementing this change. See comments. Also, I ensured that every
test that checks failures on the addAlias method gives some Grain to the
alias first, so as to flush out any corrupted-state type bugs.

Test plan: Review and `yarn test`.

[review suggestion]: https://github.com/sourcecred/sourcecred/pull/1946#pullrequestreview-446834276