Generate initial ledger for legacy projects
This adds a script for generating an initial ledger for legacy
SourceCred projects.

Usage: `yarn createLedger <legacy-instance-dir>`

Loads the project.json file from the given instance directory,
instantiates a Ledger, and fills it with users and aliases based on the
identities specified in the project.json. Prints the resulting eventLog
to stdout.

The users and their aliases are created in the same order as they are
defined in the project.json.

This is a starting point for the script; soon we will add support for
including Grain distributions and transfers.

Test plan: Run the createLedger command on the SourceCred cred instance
and check if the generated eventLog is correct