Wrap Ledger actions in an Event wrapper
The LedgerEvent wrapper contains a single action, a version string, and
a ledgerTimestamp. This way we don't need to set the ledgerTimestamp on
each action individually.

In the future, we may consider letting events hold multiple actions. I
haven't done this for now since we don't have a way of applying groups
of actions transactionally (i.e. with clean rollbacks if some actions
fail), so I feel putting multiple actions in an event would imply a
contract that we aren't well equipped to support.

Test plan: This changes the implementation but not the interfaces (since
Events are not exported). Tests have been updated.

Fixes #1988.