add `newIdentity` method
This moves responsibility for constructing identity objects out of the
ledger and into the identity module itself. This will set up future
improvements like having the identity own a subtype and directly own its
address.

This adds one new failure mode for the ledger: if it is provided a
CREATE_IDENTITY operation with an identity that already has aliases,
that should fail. I've added a test for this case.

Test plan: `yarn test` passes; coverage on the ledger is still 100%.

Part of work on #1941.