Add identity subtypes
This commit gives identities a subtype, which is one of: "USER", "BOT",
"ORGANIZATION", or "PROJECT". This gives us a lot more flexibility and
reflects the fact that there are many different kinds of identities that
might be earning Cred and receiving Grain.

The identity subtype is included in the identity address, so it will be
possible to use graph node types and graph address filtering to
distinguish them.

This is a material step towards #1941.

Test plan: Unit tests updated; `yarn test` passes.