Use ledger identities in Cred generation (#1997)
This commit:
- Adds an Identity plugin declaration in ledger/identity
(it's not really a plugin, but... close enough for now; we'll be
refactoring how plugin declarations work in the future)
- Adds a method in the identity module for producing a graph contraction
- Updates the score command so that it incorporates the identity plugin
declaration, and the node contractions

The ledger is loaded from `data/ledger.json`, assuming that file is
present.

Test plan: Manual testing on SourceCred's own instance. We'll have
automated testing after beta, when I build the DiskInstance class.