implement UserEdit component
This component allows usernames to be modified and aliases to be added.
Aliases that are already linked in the ledger are displayed read-only in
a "Synchronized Aliases" list.

Linked aliases are not yet filtered out of the autocomplete dropdown.
This should be done before initial release

test-plan: edit an existing user and attempt save the user with newly
linked alias(es) and a modified username