ledger: remove removeAlias
The `removeAlias` method will eventually be needed, since we will start
to see fraudulently/accidentally linked accounts. However, I want to
punt on implementing support until we've seen it in practice and can see
what kind of fix will be most appropriate. For the time being, in the
event of fraud it will be possible to de-activate the attacker's grain
account, thus removing their ability to earn or transfer Grain, while we
find the right solution.

Test plan: Just a removal. `yarn test`.