Ledger: only identities have grain accounts (#2004)
This commit changes the contract for the Ledger so that only identities
may have Grain accounts. I.e. regular Node Addresses may never receive
Grain. As a design decision, this is motivated by deciding that all
users should need to explicitly opt-in to recieving Grain (a future
feature). Since the opt-in will be associated with identities, we can
then remove the ability for non-identity addresses to hold Grain at all.

This turns out to massively simplify the implementation, and allows us
to not think about many otherwise-troublesome edge cases. I'm quite
happy with the cleanup.

This commit:
- modifies the APIs so that Grain Accounts are all owned by identities
- removes support for distributing Grain, pendinug a re-write of the
grain allocation module to also use identities.
- removes support for `retroactivePaid` tracking. I intend to just
remove `removeAlias` support for the time being, and add it back in when
we actually need it.