ledger: remove removeAlias (#2005)
We'll eventually need a way to unlink an alias from an account, e.g. if
the alias was fraudulently connected. But I'm not sure that removeAlias
is the right API; maybe we'll really need transferAlias. For now, I'm removing
it, so we can figure out the right API when we know the design space a little better.

For the time being, in the
event of fraud it will be possible to de-activate the attacker's grain
account, thus removing their ability to earn or transfer Grain.

Test plan: Just a removal. yarn test.