remove the alias plugin
This commit removes the alias plugin (which is being merged into core).
The only piece we keep is the actual logic for decoding aliased node
addresses, I am going to keep this temporarily around in the CLI so we
can make a reasonable WIP system for specifying identities before the
admin UI is ready.

I also removed the analysis/output module, because it depended on the
aliases, and because it is deprecated.

Test plan: Mostly just a removal, `yarn test` passes.