conditionally render react admin resources
resources should only accessible in the frontend when
`sourcecred/inititatives` are present in the instance's
sourcecred.json `bundledPlugins` array

test-plan: first build and browse the dev site while
`sourcecred/inititatives` is absent form the `bundledPlugins` array.
Ensure only the explorer is visible in the menu.

second add `sourcecred/initiatives` to the bundled plugins array and
rebuild the dev site. The dummy resource should now be visible