ledger: include identities in accounts (#2012)
The ledger module used to separately track identities and accounts. Now
each account has an associated identity. This removes a class of desync
issues, and more importantly, allows us to get all the info needed for
Grain distributions out of the ledger with a single API call.

(When distributing Grain, we need to know whether the account is active
and how much it's been paid, AND its address.)

Test plan: Unit tests updated, `yarn test` passes.