run cred explorer and react admin concurrently from dev server (#2014)
* add temporary routing for disjoint apps

Currently, the Initiatives editor is under heavy development and neither
the shape of its state, nor its interface with the sourcecred core, is
clearly defined just yet.

In order to get the Initiatives Editor merged in its current form, it
will be hidden behind the `/editor` URI while the new Explorer shall be
visible at the root of the dev server path.

test plan: run `yarn start --instance <instance>` and then browse to
both `localhost:8080` and localhost:8080/editor and ensure the Explorer
and Initiatives Editor load, in that respective order

* pass basename into router

In order to utilize the Cred Explorer (and eventually the editor) as a
subpath, react router needs knowledge of the basename (i.e. the
    application route) so it doesn't attempt to modify parts of the path
it shouldn't have access to.

Previously, this was accomplished using the home-rolled
`RelativeHistory` component in Router V3. However, in Router V4+, the
high-level router components we want to use don't expose any way to
utilize a custom history, but they do expose a pathname property, which
appears to accomplish the same thing.

test plan: build these changes using `yarn build` and copy the build
output into a local example-instance both at root where needed and in
the site directory, and then run these commands:
```bash
alias serve="python -m SimpleHttpServer 6060"
cd example-instance && serve  # then localhost:6060 should show a working frontend
cd .. && serve # then localhost:6060/example-instance should show a working frontend
```
This commit does not ensure the initiatives editor builds correctly for
production yet. This commit merely ensures the initiatives editor's
inclusion in the codebase for development does not clobber the Explorer
either in production or development builds

Secondly, the `npx serve` command does not currently work in the
example-instance with or without these changes and needs to be examined
further

* add bare react admin components

This ensures the cred explorer and react admin are running concurrently
from the dev server and can be accesses by distinct routes

test plan: start the dev server, browse to localhost:xxxx and observe
the cred explorer, browse to localhost:xxxx/editor and observe the react
admin splash page (nothing functional at this point)

* remove all zombie code

only the admin component is needed

test plan: browse to the the cred explorer and the admin app and ensure
both are visible

* rename /editor route to /admin

 genericize this subroute

 test plan: ensure the splash page is still visible at
 localhost:xxxx/admin