remove extraneous logout and refresh buttons
In order to customize the top bar, the topbar must be implemented and
modified in our codebase. The conditionally-rendered logout icon has
also been removed from the left side menu

Because we haven't made any decisions about styling, I've left the
unused styling props in, but commented out so they comply with the
linter. This way, it will be straightforward to propagate any themes if
we decide to utilize the material ui API and class system

test plan: run the dev web environment and ensure the cred explorer
renders with some data from the example instance