Chore: combine frontends within React Admin (#1989)
* unite frontends behind the react admin menu

The cred explorer and other apps are now accessible from
within react admin.

test plan: manual UAT: ensure the cred explorer can recalculate weights
Also ensure both the explorer and dummy resource are visible in the
left menu bar.

* remove path filter and note

because we're uniting everything under react admin, we don't need this
hack anymore

* implement hash routing with react admin

Hash routing allows us to utilize deeplinks and path modifications
where the path is only adjusted after the `#` symbol, so no page
loading is needed

test plan: ensure the example instance website is navigable when
served via its root and its parent folder

* Give route and menu priority to cred explorer

cred explorer is the default route and now appears on top of the
left-hand menu in the ui

* provide some padding to the cred explorer

cred explorer is super close to the margins. adding padding to the sides
and bottom of the div containing app will make it look less cramped

* load bundledPlugins into frontend

bundled plugins are needed to see what plugins should be activated in
the frontend, namely initiatives

test plan: ensure this load function doesn't fail or error on render

* conditionally render react admin resources

resources should only accessible in the frontend when
`sourcecred/inititatives` are present in the instance's
sourcecred.json `bundledPlugins` array

test-plan: first build and browse the dev site while
`sourcecred/inititatives` is absent form the `bundledPlugins` array.
Ensure only the explorer is visible in the menu.

second add `sourcecred/initiatives` to the bundled plugins array and
rebuild the dev site. The dummy resource should now be visible

* migrate load logic to its own file

test plan: ensure no load errors on render in initiatives or cred
explorer when building and browsing the dev site

* utilize the Explorer component directly

initially, the entire explorer app was utilized. This caused redundant
loading requests. and bloat. The ExplorerApp component will be retained
if the explorer is desire to be rendered as a standalone app.

test plan: ensure the explorer renders with the cred data when browsing
the dev site

* add react-redux flow libdefs

* make explorer width responsive

thanks to @hammadj for suggesting this. The only wrinkle is the
cred-over-time charts don't render well after a certain point, so these
should only render at or above a specific aspect

test plan: resize the window to see how effectively the explorer renders
at various aspects

* remove extraneous logout and refresh buttons

In order to customize the top bar, the topbar must be implemented and
modified in our codebase. The conditionally-rendered logout icon has
also been removed from the left side menu

Because we haven't made any decisions about styling, I've left the
unused styling props in, but commented out so they comply with the
linter. This way, it will be straightforward to propagate any themes if
we decide to utilize the material ui API and class system

test plan: run the dev web environment and ensure the cred explorer
renders with some data from the example instance