implement ledger identity editor POC
The ledger identity editor enables users to collect grain by allowing
them to claim aliases from various plugins and unite them under a
single identity.

This editor allows for the creation and modification of identity
usernames, along with the assignment of aliases.

Note: aliases cannot yet be unassigned as there is no graceful way to
accomplish this in the backend yet.

test plan:
 - run the dev web server and browse to the ledger admin view.
 - Try to create an identity, then assign a couple aliases to it.
 - Click on the newly created identity name in the alias list to
 reset the view and create a new identity
 - Create a second identity and assign a couple more aliases to it.
 - Note that assigned aliases are no longer visible in the dropdown
 - click between users and ensure the name and alias lists repopulate
 with the correct information