strip all over-time data, except for users' cred
This replaces our "compressByThreshold" strategy (where we include only
over-time data for nodes/flows with above 10 Cred). Instead, we keep
over time data only for user nodes. This basically gives us
CredRank-esque semantics upfront, which is nice future proofing.

Test plan: I didn't write tests because this is the sort of thing which
will either work or be obviously broken, and it's in a part of the
codebase which we expect to replace wholesale anyway (when CredRank
rolls around).

Will test via downstream usage before merging.