remove Ledger action versioning
This removes the version field from the ledger actions. Since writing
the Ledger, I've made many breaking changes. (Which has been fine, since
we don't have any ledgers in production yet.)  Over the course of making
those breaking changes, I've noticed that bumping a version on an
individual action never has felt like a useful affordance. Most of the
changes affect the semantics of the system as a whole; but the
individual primitives are quite stable.

We're more likely to need versioning on the whole Ledger, with upgrade
paths, than we are to need versioning on individual actions. And if we
do need to make a new version of a specific action, we can just put a
version suffix in the type field.

Test plan: `yarn test` suffices.